requirements.txt
setuptools~=44.0.0
PyYAML~=5.4.1
Pillow~=8.2.0
torch~=1.8.1+cu111
matplotlib~=3.4.2
numpy~=1.19.5
opencv-python~=4.2.0.32
imageio~=2.9.0
torchvision~=0.9.1+cu111
scipy~=1.4.1
vispy~=0.8.0
scikit-image~=0.18.2
tqdm~=4.61.1
transforms3d~=0.3.1
moviepy~=1.0.2
requests~=2.25.1
tensorflow~=2.6.0